Add tcl to build dependencies on Debian/Ubuntu
Before this change, a bare `make` fails with an error that `tclsh` is
not installed on Debian Stable (buster). The `tcl` package is not pulled
in by the two existing steps.

After this change `tcl` is installed explicitly and the `make` command
does not exit with an error that `tclsh` is not installed.